requirements.txt
aiogram
alembic
asyncpg
cachetools
fastapi
google_api_python_client
openpyxl
pandas
pydantic
pydantic_settings
slowapi
uvicorn
